+++ Sources/BrazeUI/InAppMessageUI/Views/InAppMessageUIContainerView.swift
@@ -64,7 +64,7 @@ extension BrazeInAppMessageUI {
       let frameEnd = window.screen.coordinateSpace.convert(keyboard.frame, to: superview)
       let frame = superview.bounds.intersection(frameEnd)
       bottomConstraint?.constant = -frame.height
-      
+
       UIView.animate(
         withDuration: 0.25,
         delay: 0,