+++ Sources/BrazeUI/InAppMessageUI/Views/InAppMessageUIModalView.swift
@@ -125,12 +125,14 @@ extension BrazeInAppMessageUI {
         bottom: padding.bottom,
         right: 0
       )
-      textContainer.layoutMargins = .init(
+
+      textView.layoutMargins = .init(
         top: 0,
         left: padding.left,
         bottom: 0,
         right: padding.right
       )
+
       buttonsContainer?.stack.layoutMargins = .init(
         top: 0,
         left: padding.left,
@@ -139,12 +141,14 @@ extension BrazeInAppMessageUI {
       )
 
       // Spacings
-      textViewStyle.headerMessageSpacing = attributes.labelsSpacing
       contentView.stack.spacing = attributes.spacing
 
-      // Fonts
-      textViewStyle.header.font = attributes.headerFont
-      textViewStyle.message.font = attributes.messageFont
+      // Text view attributes
+      textView.attributes = .init(
+        modal: message,
+        attributes: attributes,
+        traitCollection: traitCollection
+      )
 
       // Corner radius
       shadowView.layer.cornerRadius = attributes.cornerRadius
@@ -198,32 +202,11 @@ extension BrazeInAppMessageUI {
       }
     }()
 
-    public lazy var textView: UITextView = {
-      let textView = UITextView()
-      // Config defaults:
-      textView.backgroundColor = .clear
-      textView.isEditable = false
-      textView.isSelectable = false
-      textView.adjustsFontForContentSizeCategory = true
-
-      // Don't allow scrolling; the textview's parent (a scrollview) will control that.
-      // This will force the textView to render its full content height,
-      // and the textViewContainer can then take that as its content and scroll it for us.
-      // (This hot tip brought to you by https://archive.is/fMUR7 and many other results.)
-      textView.isScrollEnabled = false
-
-      // Layout defaults:
-      textView.setContentCompressionResistancePriority(.required, for: .vertical)
-      textView.setContentHuggingPriority(.required, for: .vertical)
-
-      return textView
-    }()
-
-    public lazy var textContainer: UIScrollView = {
-      let container = UIScrollView()
-      container.addSubview(textView)
-      return container
-    }()
+    public lazy var textView: ModalTextView = ModalTextView(
+      modal: message,
+      attributes: attributes,
+      traitCollection: traitCollection
+    )
 
     public lazy var buttonsContainer: StackView? = {
       let container = StackView(
@@ -256,7 +239,7 @@ extension BrazeInAppMessageUI {
       let view = StackView(
         arrangedSubviews: [
           graphicView,
-          textContainer,
+          textView,
           buttonsContainer,
         ]
         .compactMap { $0 }
@@ -294,18 +277,6 @@ extension BrazeInAppMessageUI {
       self.gifViewProvider = gifViewProvider
       self.presented = presented
 
-      self.textViewStyle = TextViewStyle(
-        header: .init(
-          color: .clear,
-          font: attributes.headerFont
-        ),
-        message: .init(
-          color: .clear,
-          font: attributes.messageFont
-        ),
-        headerMessageSpacing: attributes.labelsSpacing
-      )
-
       super.init(frame: .zero)
 
       addSubview(shadowView)
@@ -331,8 +302,11 @@ extension BrazeInAppMessageUI {
     public var theme: Braze.InAppMessage.Theme { message.theme(for: traitCollection) }
 
     open func applyTheme() {
-      textViewStyle.header.color = theme.headerTextColor.uiColor
-      textViewStyle.message.color = theme.textColor.uiColor
+      textView.attributes = .init(
+        modal: message,
+        attributes: attributes,
+        traitCollection: traitCollection
+      )
       closeButton.setTitleColor(theme.closeButtonColor.uiColor, for: .normal)
       contentView.backgroundColor = theme.backgroundColor.uiColor
       shadowView.alpha = theme.backgroundColor.a
@@ -383,19 +357,6 @@ extension BrazeInAppMessageUI {
         default:
           break
         }
-
-        // Anchor the text view's layout using margins from its parent container:
-        textView.anchors.leading.equal(textContainer.anchors.leadingMargin)
-        textView.anchors.trailing.equal(textContainer.anchors.trailingMargin)
-        textView.anchors.top.equal(textContainer.anchors.topMargin)
-        textView.anchors.bottom.equal(textContainer.anchors.bottomMargin)
-        textView.anchors.width.equal(textContainer.layoutMarginsGuide.anchors.width)
-        // Add a not-required priority to allow the text container scrollview to shrink
-        textView.anchors.height.lessThanOrEqual(textContainer.layoutMarginsGuide.anchors.height)
-          .priority =
-          .defaultHigh
-        let textViewHeightConstraint = textView.anchors.height.equal(textContainer.anchors.height)
-        textViewHeightConstraint.priority = .defaultHigh - 1
 
         // Close button
         closeButton.anchors.height.equal(closeButton.anchors.width)
@@ -509,66 +470,6 @@ extension BrazeInAppMessageUI {
       dismiss()
     }
 
-    // MARK: - Text view style
-
-    private var textViewStyle: TextViewStyle {
-      didSet {
-        if oldValue != textViewStyle {
-          self.updateTextViewContent()
-        }
-      }
-    }
-
-    private func updateTextViewContent() {
-      let textViewText = NSMutableAttributedString()
-
-      textViewText.append(
-        message.header.attributed(
-          with: [
-            NSAttributedString.Key.font: textViewStyle.header.font,
-            NSAttributedString.Key.foregroundColor: textViewStyle.header.color,
-          ],
-          {
-            $0.lineSpacing = 2 * TextViewLayoutConstants.headerLineSpacingScaleFactor
-            $0.alignment = message.headerTextAlignment.nsTextAlignment(forTraits: traitCollection)
-            $0.paragraphSpacing = max(
-              0.0,
-              textViewStyle.headerMessageSpacing
-                - TextViewLayoutConstants.headerMessageSpacingOffset)
-          })
-      )
-      // Users don't add newlines to the end of their header text
-      // Insert a newline between header and message, but not with the possible extra styling of the header (e.g. font size, etc.)
-      // The paragraph spacing set above will occur between the header text and this linebreak.
-      textViewText.append(NSAttributedString(string: "\n"))
-
-      textViewText.append(
-        message.message.attributed(
-          with: [
-            NSAttributedString.Key.font: textViewStyle.message.font,
-            NSAttributedString.Key.foregroundColor: textViewStyle.message.color,
-          ],
-          {
-            $0.lineSpacing = 4 * TextViewLayoutConstants.messageLineSpacingScaleFactor
-            $0.alignment = message.messageTextAlignment.nsTextAlignment(forTraits: traitCollection)
-          })
-      )
-
-      textView.attributedText = textViewText
-    }
-
-    private enum TextViewLayoutConstants {
-      // Manually-tuned values to get us close to our previous StackView+Label appearance.
-
-      // Soak up some vertical space that UITextView leaves above and below its text:
-      static let textContainerLayoutMargins = UIEdgeInsets(top: -8, left: 0, bottom: -8, right: 0)
-      // Scale factors for label → textview line spacing:
-      static let headerLineSpacingScaleFactor = 0.78
-      static let messageLineSpacingScaleFactor = 0.47
-      // Subtraction offset between header and message:
-      // (textview/TextKit renders a tiny bit of extra ascender+descender space that we want to eat up)
-      static let headerMessageSpacingOffset: Double = 1.0
-    }
   }
 
 }
@@ -630,7 +531,7 @@ extension BrazeInAppMessageUI {
 
       ModalView(message: .mockLong, presented: true)
         .preview(center: .required)
-        .frame(maxHeight: 375)
+        .frame(maxHeight: 700)
         .previewDisplayName("Var. | Long (constrained)")
 
       ModalView(message: .mockTallCharacters, presented: true)