+++ Sources/BrazeUICompat/ABKContentCards/AppboyContentCards.h
@@ -7,3 +7,5 @@
 #import "ABKBaseContentCardCell.h"
 #import "ABKCaptionedImageContentCardCell.h"
 #import "ABKClassicContentCardCell.h"
+#import "ABKClassicImageContentCardCell.h"
+#import "ABKControlTableViewCell.h"